Turned old-style c-Cast to static_cast

+++ src/activationFunctions/activationSoftmax.cpp
@@ -21,15 +21,15 @@ void activationSoftmax::backward(const dMatrix &dValues)
   m_dInput = nc::empty_like(dValues);
   dMatrix values = dValues.copy();
 
-  for (int32_t i = 0; i < int32_t(m_output.numRows()); i++) {
+  for (uint32_t i = 0; i < m_output.numRows(); i++) {
     //Flatten output Array
     dMatrix singleOutput = m_output.row(i).reshape(-1, 1);
     //Calculate Jacobi Matrix of Output
     dMatrix jacobianMatrix = nc::diagflat(singleOutput) - singleOutput.dot(singleOutput.transpose());
     //Sample-wise gradient
     dMatrix outputVector = values.row(i).dot(jacobianMatrix);
-    for (int32_t j = 0; j < int32_t(outputVector.numCols()); j++) {
-      m_dInput.put(i, j, outputVector.at(0, j));
+    for (int32_t j = 0; j < static_cast<int32_t>(outputVector.numCols()); j++) {
+      m_dInput.put(static_cast<int32_t>(i), j, outputVector.at(0, j));
     }
   }
 }
